+++ packages/ui/src/cairo/AccountControls.svelte
@@ -87,4 +87,4 @@
 
 <MacrosSection bind:macros={opts.macros} />
 
-<InfoSection bind:info={opts.info} />
+<InfoSection bind:info={opts.info} {errors} />