+++ packages/ui/src/confidential/ERC7984Controls.svelte
@@ -146,4 +146,4 @@
   </div>
 </ExpandableToggleRadio>
 
-<InfoSection bind:info={opts.info} />
+<InfoSection bind:info={opts.info} {errors} />