+++ packages/ui/src/stellar/StablecoinControls.svelte
@@ -116,4 +116,4 @@
 
 <TraitImplementationSection bind:explicitImplementations={opts.explicitImplementations} />
 
-<InfoSection bind:info={opts.info} />
+<InfoSection bind:info={opts.info} {errors} />